Create Game.md
guess the number game . in this game the user enter a number if number is small then it show no. is too low if enter number is larger than the number then it show its too high
and if number == input number then its print you guessed it right and it show that in how much number you try to guessed the correct number.